producer/requirements.txt
quixstreams==3.14.1
python-dotenv
zstandard

# tqdm